ecomprj/requirements.txt
asgiref==3.8.1
boto3
botocore
Django==4.2.11
dj-database-url
django-auto-logout
django-modeltranslation
django-ckeditor-5
django-crispy-forms
crispy-bootstrap5
django-dotenv
django-environ
django-recaptcha
django-filter
django-formset-js-improved
django-jazzmin
django-jquery-js
django-js-asset
django-mailgun-provider
django-plaintext-password
django-rest-auth
djangorestframework==3.15.1
djangorestframework-simplejwt
djoser
Pillow
python-decouple
rjsmin
s3transfer
validate-email
whitenoise
shortuuid
django-mathfilters
django-admin-sortable2
stripe
psycopg2-binary
paypal-payouts-sdk
paypalhttp
django-paypal
django-anymail
geoip2
requests
django-user-agents
gunicorn
channels
daphne
django-heroku
django-storages
django-taggit
django-import-export
django-templated-mail
django-tinymce
django-widget-tweaks
django-social-share
django-static-fontawesome
pandas
redis
